% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break.of.season.R
\name{break.of.season}
\alias{break.of.season}
\title{Run the break of season}
\usage{
break.of.season(r, year, mm1 = 25, mm2 = 5, date1 = as.Date(paste("2504",
  year, sep = ""), "\%d\%m\%Y"), date2 = as.Date(paste("0506", year, sep =
  ""), "\%d\%m\%Y"))
}
\arguments{
\item{r}{Data frame containing daily climate data.}

\item{year}{Year we are considering.}
}
\value{
Date of break of season.
}
\description{
Follows a simple 2-pert rule: break occurs if there is 25mm of rainfall over three 
days after 25 April, or if there is 5mm over 3 days after 5 June
}
