\docType{package}
\name{geotopbricks-package}
\alias{geotobricks}
\alias{geotopbricks-package}
\title{geotopbricks: Analyzes raster maps as input/output files from the Hydrological Distributed Model GEOtop}
\description{
  It contains implementations of parametric formulas for
  soil water retention or conductivity curve. At the
  moment, only Van Genuchten (for soil water retention
  curve) and Mualem (for hydraulic conductivity) were
  implemented. See reference
  \url{http://en.wikipedia.org/wiki/Water_retention_curve}.
  Any contributions or suggestions are warmly welcomed
}
\details{
  \tabular{ll}{ Package: \tab soilwater\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2012-11-13\cr
  License: \tab GPL (>= 2)

  \cr LazyLoad: \tab yes\cr Suggests: raster \cr }
}
\note{
  geotobricks is a on-going project. All criticism,
  comments and suggestions are well welcomed.

  geotobricks is free software: you can redistribute it
  and/or modify it under the terms of the GNU General
  Public License as published by the Free Software
  Foundation, either version 3 of the License, or (at your
  option) any later version.

  geotopbricks is distributed in the hope that it will be
  useful, but WITHOUT ANY WARRANTY; without even the
  implied warranty of MERCHANTABILITY or FITNESS FOR A
  PARTICULAR PURPOSE.  See the GNU General Public License
  for more details.

  You should have received a copy of the GNU General Public
  License along with this program.  If not, see
  \url{http://www.gnu.org/licenses/}.
}
\author{
  Emanuele Cordano \email{emanuele.cordano@gmail.org},
  Fabio Zottele \email{fabio.zottele@fmach.it}
}
\references{
  \url{wwww.geotop.org}
}
\keyword{GIS,}
\keyword{hydrology}
\keyword{package,}

