% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.water.R
\name{soil.water}
\alias{soil.water}
\title{Run fallow soil water model}
\usage{
soil.water(climate.data, soil.init, season.break = NULL,
  wheat.et.table = NULL)
}
\arguments{
\item{climate.data}{Data frame containing daily climate data.}

\item{soil.init}{Data frame containing soil data.}
}
\value{
Vector of daily plant available soil water.
}
\description{
The water balance calculation method is derived from Ritchie, J. T. 
(1972). Model for predicting evaporation from a crop with incomplete 
cover. Water Resour. Res., 8, 1204?1213
}
