\docType{package}
\name{soilwater-package}
\alias{soilwater}
\alias{soilwater-package}
\title{soilwater: Implements parametric  formulas for soil water retention or conductivity curve}
\description{
It contains implementations of parametric formulas for soil
water retention or conductivity curve. At the moment, only
Van Genuchten (for soil water retention curve) and Mualem
(for hydraulic conductivity) were implemented. See
reference
\url{http://en.wikipedia.org/wiki/Water_retention_curve}.
Any contributions or suggestions are warmly welcomed
}
\details{
\tabular{ll}{ Package: \tab soilwater\cr Type: \tab
Package\cr Version: \tab 1.0\cr Date: \tab 2012-12-4\cr
License: \tab GPL (>= 2)

\cr LazyLoad: \tab yes\cr Suggests: raster \cr }
}
\note{
soilwater is a on-going project. All criticism, comments
and suggestions are well welcomed.

Other similar packages (with similar names but different
projects) are availeble here:
http://soilwater.r-forge.r-project.org/

soilwater is free software: you can redistribute it and/or
modify it under the terms of the GNU General Public License
as published by the Free Software Foundation, either
version 3 of the License, or (at your option) any later
version.

soilwater is distributed in the hope that it will be
useful, but WITHOUT ANY WARRANTY; without even the implied
warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
PURPOSE.  See the GNU General Public License for more
details.

You should have received a copy of the GNU General Public
License along with this program.  If not, see
\url{http://www.gnu.org/licenses/}.
}
\author{
Emanuele Cordano \email{emanuele.cordano@gmail.org},
Daniele Andreis, Fabio Zottele
\email{fabio.zottele@fmach.it}
}
\references{
\url{NULL}
}
\keyword{GIS,}
\keyword{hydrology}
\keyword{package,}

