% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_balance.R
\name{water_balance}
\alias{water_balance}
\title{Water Balance}
\usage{
water_balance(data0, awc, start_month)
}
\arguments{
\item{data0}{A data.frame with three columns: month, evapotranspiration (et0) and precipitation (p)}

\item{awc}{Available water capacity}

\item{start_month}{Month of beginning of the water balance}
}
\description{
Water Balance
}
\examples{
\dontrun{
data0 <- tibble::tribble(
~month,	 ~et0,   ~p,
"Jan",	117,	275,
"Feb",	102,	218,
"Mar",	104,	180,
"Apr",	 79,	 60,
"May",	 60,	 25,
"Jun",	 49,	 20,
"Jul",	 54,	 15,
"Aug",	 74,   12,
"Sep",	 93,	 48,
"Oct",	107,	113,
"Nov",	108,	180,
"Dec",	117,	245)
water_balance(data0, awc=125, start_month="Mar")
}
}
